requirements.txt
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
bidict==0.22.0
click==8.1.3
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SocketIO==4.3.1
flaskwebgui==0.3.5
fonttools==4.31.2
h11==0.13.0
imageio==2.16.1
importlib-metadata==4.11.4
ipython==8.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.4.1
llvmlite==0.38.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.21.5
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.28
psutil==5.9.0
pure-eval==0.2.2
pycodestyle==2.8.0
pygad==2.16.3
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-engineio==3.13.2
python-socketio==4.6.0
pywin32==303
pyzmq==22.3.0
simple-websocket==0.5.2
six==1.16.0
stack-data==0.2.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
Werkzeug==2.0.2
whichcraft==0.6.1
wsproto==1.1.0
zipp==3.8.0
